Unify docs and other pages into the same PageLayout
