+++ docs/components/PageLayout.tsx
@@ -4,6 +4,8 @@ import { Toaster } from "react-hot-toast";
 import { SecondaryNav, SecondaryNavLink } from "./SecondaryNav";
 import classNames from "classnames";
 
+import { ResponsiveBreakpoint } from "../utils/style-constants";
+
 export const PageLayout = ({
   children,
   secondaryNavLinks,
@@ -16,16 +18,16 @@ export const PageLayout = ({
       <div className="page">
         <Header />
         <div
-          className={classNames("content", {
+          className={classNames("wrapper", {
             "with-secondary-nav": secondaryNavLinks,
           })}
         >
           {secondaryNavLinks && (
-            <div>
+            <div className="secondary-nav-container">
               <SecondaryNav links={secondaryNavLinks} />
             </div>
           )}
-          <div>{children}</div>
+          <div className="content">{children}</div>
         </div>
         <Footer />
       </div>
@@ -37,7 +39,7 @@ export const PageLayout = ({
           grid-template-rows: max-content 1fr max-content;
         }
 
-        .content {
+        .wrapper {
           width: 100%;
           max-width: 60rem;
           margin: 0 auto;
@@ -45,7 +47,7 @@ export const PageLayout = ({
           padding-bottom: 3rem;
         }
 
-        .content.with-secondary-nav {
+        .wrapper.with-secondary-nav {
           padding-left: 0.75rem;
           padding-right: 1.5rem;
           padding-top: 2rem;
@@ -54,6 +56,19 @@ export const PageLayout = ({
           display: grid;
           grid-template-columns: 10rem 1fr;
           grid-column-gap: 5rem;
+        }
+
+        .content {
+          min-height: 90vh; // To push footer to the bottom of the page.
+        }
+
+        @media (max-width: ${ResponsiveBreakpoint.medium}) {
+          .wrapper.with-secondary-nav {
+            grid-template-columns: 1fr;
+            grid-template-rows: 0 auto;
+            padding: 1.5rem;
+            padding-bottom: 3rem;
+          }
         }
       `}</style>
     </>
